buttons in product management view are now highlighted when pressed. there is now a confirmation window before deleting a product

+++ src/main/java/view/DeleteProductView.java
@@ -1,9 +1,13 @@
 package view;
 
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.Dialog;
 import javafx.scene.control.TextField;
 
 import java.io.IOException;
+import java.util.Optional;
 
 public class DeleteProductView {
     private MainApp mainApp;
@@ -15,7 +19,20 @@ public class DeleteProductView {
     @FXML
     private void deleteProduct(){
         try {
-            this.mainApp.getEngine().productDao().deleteProduct(productBarcode.getText());
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Vahvistus");
+            alert.setHeaderText("Vahvistus");
+            alert.setContentText("Poistetaanko varmasti?");
+
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.get() == ButtonType.OK){
+                this.mainApp.getEngine().productDao().deleteProduct(productBarcode.getText());
+                alert.close();
+            } else {
+                alert.close();
+            }
+
+
         }catch (Exception e){
             System.out.println("There was an error");
             e.printStackTrace();
